Put voter apart
